+++ FIRMWARE/plataform_tact/include/motor.h
@@ -9,5 +9,5 @@ void motor_init(void);
 void motor_set_speed(int motor_id, int speed_ms);
 uint16_t motor_get_speed(int motor_id);
 void motor_set_pos(int motor_id, int pos_um);
-uint16_t motor_get_pos(int motor_id);
+int motor_get_pos(int motor_id);
 void motor_set_pos_home(int motor_id);
